requirements.txt
ccxt==4.3.92
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
pyyaml==6.0.2
feedparser==6.0.11
yfinance==0.2.43
snscrape==0.7.0.20230622
openai==1.46.0
schedule==1.2.0